requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.3.0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
aws-cdk.aws-apigatewayv2-alpha==2.3.0.a0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha==2.3.0a0
    # via -r requirements.in
aws-cdk.aws-lambda-python-alpha==2.3.0a0
    # via -r requirements.in
cattrs==1.8.0
    # via jsii
constructs==10.0.13
    # via
    #   -r requirements.in
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
jsii==1.49.0
    # via
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   constructs
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typing-extensions==4.0.1
    # via jsii
